# speed up Relate/Contains with an rstar backed edge set intersector