--- a/geo/CHANGES.md
+++ b/geo/CHANGES.md
@@ -3,6 +3,10 @@
 
 ## Unreleased
 
+* POSSIBLY BREAKING: `GeoFloat` types must now implement `num_traits::Signed` and `num_traits::Bounded`. This shouldn't 
+  affect you if you are using a standard `Geometry<f64>` or `Geometry<f32>` or `geo::GeoFloat` generically.
+* Speed up `Relate` and `Contains` traits for large LineStrings and Polygons by using an RTree to more efficiently
+  inspect edges in our topology graph.
 * Speed up intersection checks by using a preliminary bbox check
 * Remove unneeded reference for `*MapCoords*` closure parameter.
   * <https://github.com/georust/geo/pull/810>