--- a/geo/src/algorithm/relate/geomgraph/index/mod.rs
+++ b/geo/src/algorithm/relate/geomgraph/index/mod.rs
@@ -1,7 +1,9 @@
 mod edge_set_intersector;
+mod rstar_edge_set_intersector;
 mod segment_intersector;
 mod simple_edge_set_intersector;
 
 pub(crate) use edge_set_intersector::EdgeSetIntersector;
+pub(crate) use rstar_edge_set_intersector::RstarEdgeSetIntersector;
 pub(crate) use segment_intersector::SegmentIntersector;
 pub(crate) use simple_edge_set_intersector::SimpleEdgeSetIntersector;