… geo-test-fixtures/fixtures/east_baton_rouge.wkt