@lnicola 2022-05-10
> Perhaps the worst case is comparing a big geometry to a small one.

Could we have a user-friendly API for that? I.e. loading a geometry into an R-tree, in order to be able to reuse it.