@michaelkirk 2022-05-10
> I might be missing something, but what's up with louisiana.geojson? We still seem to use the WKT version.

Thank you for catching this @lnicola. I've amended the commit to remove it.