@michaelkirk 2022-05-10
> Could we have a user-friendly API for that? I.e. loading a geometry into an R-tree, in order to be able to reuse it.

I think @urschrei's proposal for some kind of pre-indexed geometry makes sense for this. I'm happy to work on that unless someone else is planning on it. I'd prefer it to be a followup PR though, as I feel this one stands as a net win on its own.